requirements.txt
Tkinter
mysql-connector-python
pandas
openpyxl